% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_arg_linter.R
\name{match_arg_linter}
\alias{match_arg_linter}
\title{match.arg automatically detect choices if used on a function argument}
\usage{
match_arg_linter()
}
\description{
match.arg automatically detect choices if used on a function argument
}
\examples{
lintr::lint(text = "match.arg(x, c('a', 'b'))", linters = match_arg_linter())

}
