% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_product_linter.R
\name{matrix_product_linter}
\alias{matrix_product_linter}
\title{Matrix product can perform fast operations without \code{apply()}}
\usage{
matrix_product_linter()
}
\description{
Matrix product can perform fast operations without \code{apply()}
}
\examples{
# lints
lintr::lint(
  text = "apply(m, 1, function(x) sum(x * y))",
  linters = matrix_product_linter()
)

}
