% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlevels_linter.R
\name{nlevels_linter}
\alias{nlevels_linter}
\title{Use nlevels where appropriate}
\usage{
nlevels_linter()
}
\description{
Use nlevels where appropriate
}
\examples{
# lint
lintr::lint(text = 'length(levels(x))', linters = nlevels_linter())
}
