% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonportable_sort_linter.R
\name{nonportable_sort_linter}
\alias{nonportable_sort_linter}
\title{\code{sort()} can produce locale-dependent outputs when used on character vectors}
\usage{
nonportable_sort_linter()
}
\description{
\code{sort()} can produce locale-dependent outputs when used on character vectors
}
\examples{
# lints
lintr::lint(text = 'sort(x)', linters = nonportable_sort_linter())

# okay
lintr::lint(text = 'sort(x, method = "radix")', linters = nonportable_sort_linter())
}
