% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_invalid_arg_linter.R
\name{rep_invalid_arg_linter}
\alias{rep_invalid_arg_linter}
\title{Warn for invalid n argument in rep}
\usage{
rep_invalid_arg_linter()
}
\description{
Warn for invalid n argument in rep
}
\examples{
# lints
lintr::lint(text = "rep('A', n = 10)", linters = rep_invalid_arg_linter())
}
