% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setequal_linter.R
\name{setequal_linter}
\alias{setequal_linter}
\title{Flag potential opportunities for \code{setequal()} use}
\usage{
setequal_linter()
}
\description{
Flag potential opportunities for \code{setequal()} use
}
\details{
Discussed in https://github.com/r-lib/lintr/issues/1763
}
\examples{
lintr::lint(text = 'identical(sort(x), sort(y))', linters = setequal_linter())
}
