% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance_linter.R
\name{tolerance_linter}
\alias{tolerance_linter}
\title{Forbid the use of round() with all.equal() and expect_equal()}
\usage{
tolerance_linter()
}
\description{
Forbid the use of round() with all.equal() and expect_equal()
}
\examples{
# lints
lintr::lint(text = 'all.equal(round(x), y)', linters = tolerance_linter())

lintr::lint(text = 'expect_equal(round(x), y)', linters = tolerance_linter())

lintr::lint(text = 'expect_identical(round(x), y)', linters = tolerance_linter())

# okay
lintr::lint(text = 'all.equal(x, y, tolerance = 1e-4)', linters = tolerance_linter())
}
