% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorized_function_linter.R
\name{vectorized_function_linter}
\alias{vectorized_function_linter}
\title{Check for vectorized functions that might be used directly}
\usage{
vectorized_function_linter()
}
\description{
Check for vectorized functions that might be used directly
}
\examples{
# lint
lintr::lint(
  text = 'vapply(x, iconv, to = "", sub = "ASCII")',
  linters = vectorized_function_linter()
)

# okay
lintr::lint(
  text = 'iconv(x, to = "", sub = "ASCII")',
  linters = vectorized_function_linter()
)
}
